requirements.txt
streamlit
pandas
plotly
kaleido
python-dateutil
numpy
openpyxl
reportlab
